Crate wasmtime_cranelift_shared

Modules

isa_builder

Structs

CompiledFunction

Compiled function: machine code body, jump table offsets, and unwind information.

CompiledFunctionMetadata

The metadata for the compiled function.

FunctionAddressMap

Metadata to translate from binary offsets back to the original location found in the wasm input.

ModuleTextBuilder

A helper structure used to assemble the final text section of an exectuable, plus unwinding information and other related details.

Relocation

A record of a relocation to perform.

Enums

RelocationTarget

Destination function. Can be either user function or some special one, like memory.grow.

Constants

ALWAYS_TRAP_CODE

A custom code with TrapCode::User which is used by always-trap shims which indicates that, as expected, the always-trapping function indeed did trap. This effectively provides a better error message as opposed to a bland “unreachable code reached”

CANNOT_ENTER_CODE

A custom code with TrapCode::User corresponding to being unable to reenter a component due to its reentrance limitations. This is used in component adapters to provide a more useful error message in such situations.

Traits

CompiledFuncEnv

Trait used in the CompiledFunction to resolve the locations of external name references in a compiled function.

Functions

clif_flags_to_wasmtime

Converts cranelift_codegen settings to the wasmtime_environ equivalent.

mach_trap_to_trap

Converts machine traps to trap information.