wasmtime_cranelift_shared

Module obj

Object file builder.

Creates ELF image based on Compilation information. The ELF contains functions and trampolines in the “.text” section. It also contains all relocation records for the linking stage. If DWARF sections exist, their content will be written as well.

The object file has symbols for each function and trampoline, as well as symbols that refer to libcalls.

The function symbol names have format “_wasm_function_N”, where N is FuncIndex. The defined wasm function symbols refer to a JIT compiled function body, the imported wasm function do not. The trampolines symbol names have format “_trampoline_N”, where N is SignatureIndex.

Structs

ModuleTextBuilder

A helper structure used to assemble the final text section of an exectuable, plus unwinding information and other related details.