Struct wasmtime_cranelift_shared::FunctionAddressMap

pub struct FunctionAddressMap {
    pub instructions: Box<[InstructionAddressMap]>,
    pub start_srcloc: FilePos,
    pub end_srcloc: FilePos,
    pub body_offset: usize,
    pub body_len: u32,
}

Metadata to translate from binary offsets back to the original location found in the wasm input.

Fields
instructions: Box<[InstructionAddressMap]>

An array of data for the instructions in this function, indicating where each instruction maps back to in the original function.

This array is sorted least-to-greatest by the code_offset field. Additionally the span of each InstructionAddressMap is implicitly the gap between it and the next item in the array.

start_srcloc: FilePos

Function’s initial offset in the source file, specified in bytes from the front of the file.

end_srcloc: FilePos

Function’s end offset in the source file, specified in bytes from the front of the file.

body_offset: usize

Generated function body offset if applicable, otherwise 0.

body_len: u32

Generated function body length.